450 box pictures
just some pictures of the box and door, if you want more I can send them, going to try and add a place at the back of the box to put the arduino, will just slide in or something. main thing on the left hand side of the box is to hold the servo. will print out and bring to class when It finishes.