Create 450_Project_arduinoCode.ino
This is just some code I originally created to test the servo, and both LED's.  I started to create the method that I think we would need to call, but I don't know yet how we will actually receive the boolean "unlock" in the method parameter from c#, we can talk about it in out meeting coming up.